rearrange presenters - separate info, menu, filter
